refactored cmake file, moved headers to include dir
