+++ figurer-project/include/externalfilescanner.h
