+++ figurer-project/include/iexternalfileparser.h
