+++ figurer-project/include/projectmanager.h
