+++ figurer-project/include/statuswidget.h
